splitted useFileMap hook, added tracing for touched and deleted files
