+++ src/Routes/Tables/DataSources/FileBrowser.js
@@ -1,10 +1,4 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { useCallback, useMemo, useReducer } from 'react';
 import PropTypes from 'prop-types';
 import {
   ChonkyActions,
@@ -17,7 +11,7 @@ import {
   setChonkyDefaults,
 } from 'chonky';
 import { ChonkyIconFA } from 'chonky-icon-fontawesome';
-import { stratify, flatten } from './stratifier';
+import useFileMap from './useFileMap';
 
 /**
  * @typedef {import('chonky').ChonkyFileActionData} ChonkyFileActionData
@@ -31,126 +25,6 @@ import { stratify, flatten } from './stratifier';
  */
 
 setChonkyDefaults({ iconComponent: ChonkyIconFA });
-
-export const useCustomFileMap = (filesList, rootFolderId = '/') => {
-  const baseMap = useMemo(() => stratify(filesList), [filesList]);
-  const [fileMap, setFileMap] = useState(baseMap);
-  const [currentFolderId, setCurrentFolderId] = useState(rootFolderId);
-
-  const retrieveFiles = useCallback(() => flatten(fileMap), [fileMap]);
-
-  const currentFolderIdRef = useRef(currentFolderId);
-
-  useEffect(() => {
-    currentFolderIdRef.current = currentFolderId;
-  }, [currentFolderId]);
-
-  const resetFileMap = useCallback(() => {
-    setFileMap(baseMap);
-    setCurrentFolderId(rootFolderId);
-  }, [baseMap, rootFolderId]);
-
-  const deleteFiles = useCallback(
-    /** @param {CustomFileData[]} files */
-    files =>
-      setFileMap(oldFileMap => {
-        const newFileMap = { ...oldFileMap };
-        files.map(file => {
-          delete newFileMap[file.id];
-          if (file.parentId) {
-            const parent = newFileMap[file.parentId];
-            const newChildrenIds = parent.childrenIds.filter(
-              /** @param {string} id */
-              id => id !== file.id
-            );
-            newFileMap[file.parentId] = {
-              ...parent,
-              childrenIds: newChildrenIds,
-              childrenCount: newChildrenIds.length,
-            };
-          }
-          return undefined;
-        });
-        return newFileMap;
-      }),
-    []
-  );
-
-  const moveFiles = useCallback(
-    /**
-     * @param {CustomFileData[]} files
-     * @param {CustomFileData} source
-     * @param {CustomFileData} destination
-     */
-    (files, source, destination) =>
-      setFileMap(oldFileMap => {
-        const newFileMap = { ...oldFileMap };
-
-        const moveFileIds = new Set(files.map(f => f.id));
-        const newSourceChildrenIds = source.childrenIds.filter(
-          id => !moveFileIds.has(id)
-        );
-        newFileMap[source.id] = {
-          ...source,
-          childrenIds: newSourceChildrenIds,
-          childrenCount: newSourceChildrenIds.length,
-        };
-        const newDestinationChildrenIds = [
-          ...destination.childrenIds,
-          ...files.map(f => f.id),
-        ];
-        newFileMap[destination.id] = {
-          ...destination,
-          childrenIds: newDestinationChildrenIds,
-          childrenCount: newDestinationChildrenIds.length,
-        };
-        files.map(file => {
-          newFileMap[file.id] = {
-            ...file,
-            parentId: destination.id,
-          };
-          return undefined;
-        });
-        return newFileMap;
-      }),
-    []
-  );
-
-  const createFolder = useCallback(
-    /** @param {string} folderName */
-    folderName =>
-      setFileMap(oldFileMap => {
-        const newFileMap = { ...oldFileMap };
-        newFileMap[folderName] = {
-          id: folderName,
-          name: folderName,
-          isDir: true,
-          modDate: new Date(),
-          parentId: currentFolderIdRef.current,
-          childrenIds: [],
-          childrenCount: 0,
-        };
-        const parent = newFileMap[currentFolderIdRef.current];
-        newFileMap[currentFolderIdRef.current] = {
-          ...parent,
-          childrenIds: [...parent.childrenIds, folderName],
-        };
-        return newFileMap;
-      }),
-    [currentFolderIdRef]
-  );
-
-  return {
-    fileMap,
-    currentFolderId,
-    setCurrentFolderId,
-    resetFileMap,
-    deleteFiles,
-    moveFiles,
-    createFolder,
-    retrieveFiles,
-  };
-};
 
 const fileActions = [ChonkyActions.CreateFolder, ChonkyActions.DeleteFiles];
 
@@ -168,9 +42,12 @@ const FileBrowser = ({ files: srcFiles }) => {
     moveFiles,
     createFolder,
     retrieveFiles,
-  } = useCustomFileMap(srcFiles);
+    touchedFiles,
+    deletedFiles,
+  } = useFileMap(srcFiles);
 
-  const [isFlat, setIsFlat] = useState(false);
+  const [isFlat, toggleFlat] = useReducer(state => !state, false);
+
   const files = useMemo(() => {
     if (isFlat) {
       return Object.values(fileMap).filter(file => !file.isDir);
@@ -194,18 +71,18 @@ const FileBrowser = ({ files: srcFiles }) => {
   }, [currentFolderId, fileMap]);
 
   const handleFileAction = useCallback(
-    /** @param {ChonkyFileActionData} data */
-    data => {
-      switch (data.id) {
+    /** @param {ChonkyFileActionData} action */
+    action => {
+      switch (action.id) {
         case ChonkyActions.DeleteFiles.id:
-          return deleteFiles(data.state.selectedFilesForAction);
+          return deleteFiles(action.state.selectedFilesForAction);
         case ChonkyActions.MoveFiles.id:
           // eslint-disable-next-line
-          const { payload } = data;
+          const { payload } = action;
           return moveFiles(payload.files, payload.source, payload.destination);
         case ChonkyActions.OpenFiles.id:
           // eslint-disable-next-line
-          const { targetFile, files } = data.payload;
+          const { targetFile, files } = action.payload;
           // eslint-disable-next-line
           const fileToOpen = targetFile ?? files[0];
           if (fileToOpen && FileHelper.isDirectory(fileToOpen))
@@ -234,7 +111,7 @@ const FileBrowser = ({ files: srcFiles }) => {
       <button type="button" onClick={() => console.log(retrieveFiles())}>
         print updated files
       </button>
-      <button type="button" onClick={() => setIsFlat(state => !state)}>
+      <button type="button" onClick={toggleFlat}>
         toggle flat
       </button>
       <div style={{ height: 400 }}>
@@ -249,6 +126,9 @@ const FileBrowser = ({ files: srcFiles }) => {
           <FileList />
           <FileContextMenu />
         </ChonkyFileBrowser>
+        <pre>{JSON.stringify(deletedFiles, null, 2)}</pre>
+        <pre>{JSON.stringify(touchedFiles, null, 2)}</pre>
+        <pre>{JSON.stringify(files, null, 2)}</pre>
         <pre>{JSON.stringify(retrieveFiles(), null, 2)}</pre>
       </div>
     </>
